requirements.txt
fastapi
uvicorn
httpx
dotenv
pyannote.audio
torch
google-genai
python-multipart